Extract Skydive Pro View side view into its own class
Move the side-projection chart (grid, trajectory, flares, max-speed
marker, crosshair + comparison tooltip, interaction) and the wind
indicators out of skydive_performance_track_controller into a
SkydivePerformanceSideView class.

The controller now feeds it data via render(), drives the cursor via
setPosition(), and receives seeks back through an onSeek callback.
Shared point math (bearing, target-index, nearest-by-player-time,
interpolate-by-player-time) moves to utils/tracks/pointHelpers and is
reused by both the side view and the map markers.

Also: drop the unused timeRange, and attach the side view's pointer
listeners once instead of on every render.

No behavior change. Controller drops from ~2560 to ~1800 lines.
